Add `colors` support for `tf.image.draw_bounding_boxes`
This fix tries to address the issue raised in 15692 where
it was not possible to specify the colors for boxes in
`tf.image.draw_bounding_boxes`. Instead, a predefined
fixed color table was used to cycle through colors.

This fix adds `colors` Input to `DrawBoundingBoxexV2`
so that it is possible to specify the color. In case
no color is specified, the default color table will
be used.

Since there is an API change, the op is labeled as V2.

This fix fixes 15692.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>