Allow tf.contrib.tpu._prepare_validation_data to modify x,y
When the user passes a validation_split parameter, the samples allocated to the val_x, val_y populations should be removed from the training data set (x, y).  In the _prepare_validation_split function, the validation split parameter branch already creates the correctly truncated x, y but does not return those modified x, y values back.  As a result, the validation samples are in both the training data set and the validation data sets. The validation loss/metrics are giving artificially good results since the model is training also the samples in the validation set.