Remove a checkpointable dependency on __delattr__
I ran into this when writing an error message for ".signatures" being a reserved attribute in tf.saved_model.save. I want to write "'del obj.signatures' if you don't care", but that didn't actually remove the dependency.

This isn't any more unsafe than overwriting an attribute with __setattr__, which we already support. It's a somewhat weird thing to do in terms of checkpointing, though.

PiperOrigin-RevId: 231853096