Fix test sharding
Test sharding appears to have been broken by commit 87cc788.  The result
being that many tests are simply not run when sharding is enabled.
The issue seems to be that the filtering for sharding is happening
twice, once in tensorflow/python/platform/googletest.py and then once
again in external/absl_py/absl/testing/absltest.py.  This commit fixes
the issue by removing the shard filtering code in googletest.py.

Fixes: https://github.com/tensorflow/tensorflow/issues/25594