Fix public api_docs generation for tensorboard and tensorflow_estimator
Switch from a single base_dir and code_url_prefix to, effectively, a map from base dir to url_prefix.

The doc tools only document objects in modules inside `base_dir`, and add a URL to the code created from the `code_url_prefix` + the object's file path relative to the `base_dir`. So with this change, the `defined_in` links will point back the github-repos for tensorboard, and tensorflow_estimator (master).

PiperOrigin-RevId: 236205402