Fix default logdir for TensorBoard Keras callback on Windows
As cl/235232824, but this one probably hits production users who do not
specify a logdir, because the default argument value is not portable.
For root cause, see:
<#26021>

Tested:
Added a regression test. On Unix, the test passes both before and after
this change. On Windows, the test fails before this change with

    tensorflow.python.framework.errors_impl.NotFoundError: Failed to create a directory: ./logs\train; No such file or directory [Op:CreateSummaryFileWriter]

and passes after this change.

RELNOTES: n/a
PiperOrigin-RevId: 236228230