Add tf.repeat support equivalent to numpy.repeat
(Most of the credit goes to RaggedTensor author,
couldn't find the GitHub handle but the related commit
is bad5d1a)

This PR tries to address the issue raised in 8246
to have tf.repeat equivalent to numpy.repeat.

Multiple attempts have been made in the past.
My previous PR 15224 add the support with C++ which
was closed, as a python implementation relying on existing
ops is more desireable.

This PR
- moves the repeat implementation in ragged_util.py
to array_ops.py so that it is possible to get exposed.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>