Fix incorrect deprecation warning for tf.numpy_function
This fix tries to address the issue raised in 26735 where
usage of tf.numpy_function in v2 generates incorrect deprecation
warning:
```
tf.py_func is deprecated in TF V2.
```

The reason was that tf.numpy_function calls tf.py_func internal
so a warning (on tf.py_func) is always in place even though
tf.numpy_function is the correct usage in v2.

This fix reorders the calling so that incorrect warning could be addressed.

This fix fixes 26735.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>