Reduce overhead in HloInstruction::PostOrderDFS
Previously we called HloModule::instruction conut every time to reserve
the visitor state what goes through all of the computations one by one to
count the number of instructions. This is problematic if we use the same
visitor to visit all of the computations within a module as it will become
quadratic in the number of computations.

This change updates the code to only reserve a state if the state is empty
what removes the quadratic behaviour. Additionally it changes the
HloModuleGroupMetadata pass to re-use a FunctionVisitor instead of creating
a new one for every computation to avoid the quadratic behaviour.

There are other passes out there what aren't reuse the visitor (most notably
AlgebraicSimplifier) but fixing them are a bit more tricky so they will be
dealt with in a followup CL.

PiperOrigin-RevId: 242448643