[tf.data] Sharing FunctionLibraryDefinition across different instances of CapturedFunction of the same op kernel.
This CL is a fix to an issue introduced by cl/240419515 which introduced per-function FunctionLibraryDefinitions. Because function handles used for function instantiation encode the address a FunctionLibraryDefinition, the unintended effect of cl/240419515 was that different instances of the same user-defined function nested under tf.data transformation such as `interleave` or `flat_map` would end up with different handles and their kernels would thus be created for every input element of `interleave` and `flat_map`. In other words, the allocated memory would be proportional to the cardinality of the input dataset.

This CL fixes the problem by introducing a shared FunctionLibraryDefinition in the op kernels that create instances of CapturedFunction.

PiperOrigin-RevId: 242782635