Support captured Datasets in tf.functions exported using tf.saved_model.save
Adds a (purely symbolic) _Retval node to the Dataset's GraphDef representation to make it clear what the output is. This doesn't seem to hurt anything, and sounds cleaner than just assuming that the "last" node in the graph is the output.

With an identified output, we can trace a function which re-creates the variant Tensor and put that function in the SavedModel (along with any functions the Dataset references). One alternative would be to put the variant Tensor in the SavedModel's checkpoint, but that is much more opaque (vs. being able to inspect the Dataset ops creating the variant Tensor in the saved_model.pb itself) and generally there's no reason to put Datasets in non-SavedModel training checkpoints.

Makes Datasets Trackable so they can attach a TrackableResource to themselves.

This change does not make Datasets usable in new code when importing a SavedModel since there's no Python wrapper for the variant Tensor. It just provides the SavedModel code with a way to create the Tensor and an object to store it in, which allows existing functions loaded from the SavedModel use it.

In the future we may want to also deal with assets somehow; right now files referenced by Datasets are not collected as assets, which means the resulting SavedModels are not hermetic.

PiperOrigin-RevId: 244103577