Exclude device_tracer for CPU build
cl/243861265 introduced a dependency on tensorflow/core/platform/default/device_tracer.cc, however this file is actually empty during the CPU build because GOOGLE_CUDA is not defined. So the cc library for device_tracer is empty, and when force linking (/WHOLEARCHIVE is used for linking shared library) an empty cc library with MSVC, we hit this compiler bug:
https://support.microsoft.com/en-hk/help/4020481/fix-link-exe-crashes-with-a-fatal-lnk1000-error-when-you-use-wholearch

This is the root cause of b/130835153 and b/130748329

Since device tracer is only for GPU build, the solution is to exclude it from CPU build.

PiperOrigin-RevId: 244872492