Enable distributed subclassed Keras models outside of graph mode or TPU.
This is done by running `Model.train_on_batch` of the model under each replica instead of cloning the model for each replica.  The per-replica function is executed with @tf.function  .  It's also cached in the model.  This approach allows to distribute Functional, Sequential and subclassed models.

At this stage, the new behavior can be enabled by passing `cloning=False` to `Model.compile`.  In addition to that, cloning is still going to be used for TPU or for graph mode.  Reasons for that:
 - graph mode might have a different priority than eager mode
 - TPU needs to be tackled separately.
 - decoupling submitting of this CL from the performance and full-feature coverage concerns.  Once its ready for the prime time we can remove the `cloning` switch.

I verified this change by running Resnet on CIFAR with and without the change:
 - without the change 1 GPU: 'accuracy_top_1': 0.9314903616905212
 - with the change 1 GPU: 'accuracy_top_1': 0.9306890964508057
 - with the change 2 GPUs: 'accuracy_top_1': 0.9356971383094788

Resnet on imagenet got "sparse_categorical_accuracy: 0.7629".  With FP32, 8*128 batch the speed was 2313 examples/sec.  With XLA it was 2782.

PiperOrigin-RevId: 245240806