Templatize Gemm in a 'quantization flavor' distinguishing per-channel vs uniform quantization.
This isn't our preference design-wise, and is unnecessary as far as ruy is concerned, as in ruy this is just a runtime switch.

But we want to avoid code size regressions also in the currently default case where gemmlowp not ruy is used, and in gemmlowp this is a compile-time switch.

Moreover, the actual instantiations that TFLite needs are such that there is a real overall binary size regression unless this is templatized in this way. That is because TFLite uses 2 combinations of { input type, quantization flavor }:
 1. { uint8, uniform quantization }
 2. { int8, per-channel quantization }
TFLite does not use other combinations such as {uint8, per-channel}, so if we force gemmlowp to instantiate this case, we regress binary size overall.

PiperOrigin-RevId: 246070679