Change GraphCycles to use a properly ordered set
This CL changes GraphCycles's representation for input and output edge sets.

 - Instead of an unordered_set we now use an explicitly ordered set
 - Under the hood we use ABSL's flat_hash_map instead of unordered_set

Finally we change GraphCycles::Predecessors and GraphCycles::Successors to not
return edge sets by copy.  Instead we return an absl::Span and explicitly make a
copy in the one case we need a copy.

This speeds up MarkForCompilation in an internal benchmark by approximately 2x.

PiperOrigin-RevId: 247114705