Read chunks of size at most INT_MAX using pread
On macs pread returns EINVAL with chunks that don't fit in a 32-bit integer.

I could write a unit test that writes and reads 4gb, but I don't particularly want to maintain it.

If this causes performance issues on other platforms, we could make this chunking logic mac-specific. I don't see a strong reason to right now.

Fixes #18769.

PiperOrigin-RevId: 247125959