Prevent clustering if it delay a device.
Before if a cluster on the CPU had two inputs, each on a different GPU, XLA would merge the CPU cluster with one of the GPU clusters. For example, suppose cluster G0 is on GPU:0, cluster G1 is on GPU:1, and cluster C0 is on CPU:0. Also suppose C0 took G0 and G1 as input. Before this change, XLA would merge C0 with G0, preventing G0 from running until G1 completed.

PiperOrigin-RevId: 247296425