Support fault-tolerance in model.fit() by ModelCheckpoint callback capable of restoring `epoch` in sync training.
One of the challenges to ensure the training continues in the event a worker fails and recovers, is the epoch information gets lost. This change makes ModelCheckpoint save the epoch to model file along with weights and other information, and load it at the beginning of training if `load_weights_on_restart` is True.

A unit test using multi-threading to simulate multi workers with preemption is provided to ensure the results with and without interruption are the same.

PiperOrigin-RevId: 247494880