Unify ScopedDeviceMemory and OwningDeviceMemory.
Previously, ScopedDeviceMemory was parameterized by a type it holds,
and was only constructible by a stream executor, while OwningDeviceMemory was
constructible from an arbitrary allocator, but was not parameterized.

This refactoring leaves a single class ScopedDeviceMemory which is parameterized,
can be constructed from an arbitrary allocator, and has helper constructors
for constructing from a stream executor.
OwningDeviceMemory is left as a typedef for ScopedDeviceMemory parameterized on
uint8 (holds arbitrary bytes, no alignment guaranteed).

PiperOrigin-RevId: 248234830