Always initialize collective group runtime details using NcclReduce when collective_nccl is set to True.
Before this change, we would look up the collective type and invoke collective
group runtime initialization based on the type.  However, if the same group
which was initialized using a non-NCCL collective was later used for a NCCL
collective, the lack of a communicator key would lead to failure.

After this change, if collective_nccl is set, then all collective groups are
initialized as if they may be used for a NCCL collective.  In the current
implementation, this is achieved by calling group initialization for
NcclReduce.  In the future, there will be a common group initialization for all
NCCL collectives, including broadcast and all-gather, which will be invoked.

The only action in NCCL group initialization is calling ncclGetUniqueId, which
should be cheap.

PiperOrigin-RevId: 248547157