[ROCm] Fix for the broken `--config=rocm` build.
This PR contains a fix for the broken `--config=rocm` build.

Currently the file `tensorflow/core/util/gpuLaunch_config.h` is not included in any code that is enabled for the `--config=rocm` build. Once that is included within ROCm enabled code, it will break the `--config=rocm` build, because that file currrently contains a couple of Cuda* names that should not be visible in the ROCm build. This commit/PR fixes that.