[XLA] Change conditional simplifier to process instructions in post order.
Previously, conditional simplifier first collects the conditional instructions
in the order they are stored in the instruction vector in HloComputation, then
processes each collected instruction. This is a problem if the process of an
instruction removes another instruction that is ordered after as dead code.
Changing the pass to collects the conditional instructions in post order can
avoid such a situation.

Add a test case.

PiperOrigin-RevId: 249324972