Disable Tensor slice assign, since it's currently broken and throwing hard-to-understand exceptions
The code path being reverted only fires if var=None, which gets converted to a Tensor and always fails.

This is a partial revert of #27327

It's relatively trivial to instead make this work (s/var/input_/), but I'm not convinced that it's a good API, since we've seen at least one person who accidentally got a Tensor here and was using it like variable slice assignment (which of course doesn't modify the Tensor in-place and so doesn't do what they were expecting). But for `var` always being None and so throwing an opaque exception their code would have silently done the wrong thing.

So if someone else wants to re-do Tensor assignment working this time I think that's fine. But as a temporary measure we should have the more informative error we were throwing before.

PiperOrigin-RevId: 249555761