Introduce a proper fix for the segfault hotfixed in cl/249335436
The actual bug was that the passed stream executor was null,
and calling `->device_ordinal()` on it expectedly NPEs.

The underlying problem is more subtle: the vector of stream executors
returned from PlatformUtil contained null entries to correspond to "not
available" devices, and Backend filters out this vector,
yet unfiltered version with nullptrs was passed to StreamExecutorMemoryAllocator.

This change simplifies semantics, so that PlatformUtil::GetStreamExecutors
always returns a vector with no nullptr entries, which can be used directly
by Backend and the StreamExecutorMemoryAllocator.

Inside tests we just use [0] index because it corresponds to the default
executor, and for tests any working device would do.

PiperOrigin-RevId: 249877692