Do not invoke ScopedAllocator when an OpKernel calls allocate_temp with a scope_id.
This change only has effect when running a TF graph/function with collective
ops and with ScopedAllocatorOptimizer enabled.

Before this change, if an OpKernel called allocate_temp with
AllocatorAttributes containing a meaningful scope id, and later called
set_output with this tensor, it would cause a double allocation from a single
use ScopedAllocatorInstance.

After this change, there are only 2 ways to cause an allocation from
ScopedAllocator:
1. allocate_output with AllocatorAttributes having a scope id.  This is the
most performant way and will not lead to extra copies.
2. set_output on an index which should have been allocated with scoped
allocator.  In this case, we check if the output has not already been allocated
and if not, we force an allocation and a copy.
All other calls to allocate_* in OpKernel will set scope_id to -1 to ensure
that we are not calling into ScopedAllocator.

The performance implication of this change is that kernels should call
allocate_output whenever possible to ensure no extra copies.  Any other call,
such as allocate_temp followed by set_output, will lead to an intermediate
memory allocation, and then a copy into the scope allocated buffer.

This change also removes calls to GetScopedAllocator inside memory tracking
branches.

PiperOrigin-RevId: 249888173