Call dataset.get_next() under the tf.function when distributing a Keras model.
This has performance benefits due to get_next() being part of the same graph as the forward pass.

Batch dimensions will be less known now. We used to synchronize getting the actual values and the forward computation, which is why the tf.function knew the exact shapes.  The batch dimension is going to be unknown inside the tf.function now, just like in the legacy graph mode.

It also decreases the amount of tf.function's re-tracings that occur in our current official Transformer model.

PiperOrigin-RevId: 250991298