Remove deleted symbols from tensorflow.__all__. Fix #29536
This fix is required as the API generation job doesn't check that `__all__` has been initialized to contain the `core`, `compiler` and `python` symbols just a little before these symbols are deleted. As a result, when someone does `from tensorflow import *`, the interpreter sees these symbols in `tensorflow.__all__` and tries to bring them up in scope. However, as they are deleted this fails.

5a93bb solved this issue for V1 API and the issue surfaced for V2 API with f1ffa02, the virtual pip package change.

PiperOrigin-RevId: 252060393