[XLA] Fix ConvertRandomBitsToUniformFloatingPoint to exclude the leading one bit in number of mantissa bits.
We use primitive_util::SignificandWidth to find the number of mantissa bits for
a floating point data type. This actually results in one more than the actual
number of mantissa bits. Need to subtract one from this value. This was a bug
we introduce when extending the routine for double precision in cr/251649864.

PiperOrigin-RevId: 252142995