[Keras] Implement a default compute_output_signature to base layer.
This should work for all existing layers, and needs to be modified for
only custom RNNs whose output dtype does not match self.dtype or input dtype.

Modified test utils to ensure this new implementation is consistent for all
core keras layers using this test harness.

Added a test to ensure this works with a mixed precision dtype policy.

PiperOrigin-RevId: 252832062