Specialize execute for internal Send/Recv ops
Send/Recv ops internally constructed due to copy operations were using
the EagerExecute call which leads to a few undesirable behaviours such
as polluting the kernel cache, user-level logging and risk of infinite
recursion.

As a fix we introduce specialized code-paths for Send and Recv which
execute the internal ops. Going forward we probably want to avoid the op
abstraction altogether which will further simplify the code.

Additionally:
- Add helper functions to share code between EagerRemoteExecute,
  ExecuteSend & ExecuteRecv.
- Clean up IS_MOBILE_PLATFORM pre-processor usage to capture complete
  function definitions.
PiperOrigin-RevId: 255346842