[ROCm] Skipping subtests that check support for the Cholesky and QR ops
ROCm platform currently does not support the Cholesky and QR ops (on the GPU)

This commit skip subtests (within python unit-tests) that test this functionality. The "skip" is guarded by the call to "is_built_with_rocm()", and hence these unit-tests will not be affected in any way when running with TF which was not built with ROCm support (i.e. `--config=rocm`)