Create a generic quantization pattern for quantization graph rewrite
This pattern assumes that the quantization parameter propagation pass has
propagated the quantization parameters to all the quantizable ops. Instead of
defining table gen pattern for each of the quantizable ops, we can use this
single pattern to match all the ops.

The tests are mainly the existing ones, plus some manual tests.

PiperOrigin-RevId: 257213495