[XLA:Python] Simplify the Device::ThenRelease() helper API.
Consolidate on the variant that calls destructors from a worker thread; of the two uses of the non-threaded version, one was buggy since it might have freed GPU buffers, and for the other the benefit seemed minimal.

Also remove overly complicated and confusing code that tried to control which thread released a reference; the only thing that matters is that on GPU references to GPU memory aren't released from a device callback.

Change PythonRefManager so the RAII class it returns is non-copyable but is wrapped in a std::shared_ptr<>. This makes its copying semantics easier to reason about.

PiperOrigin-RevId: 257597653