[ROCm] Fix for the broken `--config=rocm` build
The following PR/commit breaks the `--config=rocm` build

https://github.com/tensorflow/tensorflow/pull/29224

It introduces a couple of GPU kernels which have CUDA specific data-structures + syntax, which leads to a compile failure on the ROCm platform.
This PR fixes the ROCm build by converting the CUDA specific code to its equivalent GPU generic code.