Fix `special_math_ops._transpose_if_necessary()` for Python 3.x.
We were comparing the list of permutation indices to `range(len(perm))`, to avoid unnecessary transposes. On Python 2.x, this is a list, which means the equality comparison has the desired effect. On Python 3.x it is a range iterator, and the equality check fails, creating unnecessary transposes for users of `tf.einsum()`.

PiperOrigin-RevId: 259131715