Sort NCCL collective participants by rank in addition to stream executor.
Before this change, if participants in a NCCL collective were on the same GPU,
their ordering was non-deterministic when launching the `ncclAllReduce`
kernels.  However, `NcclManager::LoopKernelLaunches` picked the `Participant`
whose index in the list of participants matched the `CommunicatorMember`'s
index.  So it was possible to associate a `Participant` with a communicator of
the incorrect rank.

The test added in this change, which launches broadcast kernels from concurrent
threads, tickles this bug.  We did not run into this bug before because the
previous concurrent tests only launch all-reduce kernels, in which all ranks
play the same role.

The fix to the issue is to sort participants on the same GPU, i.e. with the
same stream executor, based on their rank, thereby ensuring that the
`Participant` <-> `CommunicatorMember` matching is correct at launch time.

PiperOrigin-RevId: 260230221