Use an expandable threadpool rather than spawning new threads in collective ops.
Collective op kernels may perform blocking work.  Before this change, we would
call `SchedClosure` with the blocking function, which spawns and detaches a new
thread.

This change adds an `UnboundedWorkQueue` in `CollectiveExecutorMgr`.
`UnboundedWorkQueue` internally maintains an expandable pool of threads.  If
all threads are busy, `UnboundedWorkQueue::RunClosure` adds a new thread to the
pool.  Over time, we expect the number of threads in collective ops to settle
on the number of concurrent executions.

On `CollectiveOpBenchmark`, this change leads to 50-60% improvement with
short-running collective ops.

PiperOrigin-RevId: 260646003