Insert TFL QDQs after TF FakeQuant op in the first greedy pattern rewrite call
In a recent change, this rewrite pattern was moved to the second greedy pattern
rewrite call, but the TF FakeQuant ops are constant folded in the first greedy
pattern rewrite call, so the narrow_range and bit_width attribute from the TF
FakeQuant ops are missing. We failed to test this case since the default values
in the following passes matches the TOCO requirement. This patch restores the
original behavior and uses the TFL quantize and dequantize ops to preserve
these information.

This patch fixed another related bug that the generated patterns are not
applied in the second greedy pattern rewrite call, so the TF transpose/reshape
are not lifting to enable the constant folding. This patch added the generated
patterns to the second greedy pattern rewrite call. At the same time, two
lifting rules were added so the TFL quantize and dequantize ops are handled.

This patch also improved the implementation of the TFL QDQs inserting pattern.
Some related tests are simplified to only check necessary invariant.

PiperOrigin-RevId: 261267567