Don't duplicate scatter HLO instructions in the D2D copy thunk
The HLO profiler expects at most one Thunk to correspond to an HLO instruction.
In case we have to emit a D2D thunk the SequentialThunk is the "main" thunk
corresponding to the scatter instruction from the perspective of the profiler.

We already do the right thing for the KernelThunk.

PiperOrigin-RevId: 261989800