Replaced #include absl/string_view.h with template magic. Added tstring operators.
The inclusion of absl::string_view precludes the use of tstring in tflite.
Given that, in order to mitigate the forced inclusion absl/strings/string_view.h while
providing convenience methods for implicit conversion, we replace explicit uses
of absl::string_view with a forward decleration and associated templates.

Additionally, we include a few string comparators currently used in TF.

PiperOrigin-RevId: 262035876