Add a `communication_hint` argument to collective ops.
Before this change, the only way for a user of collective ops to specify a
particular collective implementation, such as ring or nccl, was by setting a
flag in `ConfigProto`.  This flag was set globally and would switch the
implementation of all collectives.

This change adds an extra argument to the op which enables specifying the
user's preferred implementation on an op-by-op basis.  The runtime treats this
argument as a hint and may fall back to a different implementation in favor of
erroring out.

This would also enable more fine-grained control and experiments by users who
may want to mix different collectives in a single program.

The default value is `auto`, which leaves the choice to the runtime.

PiperOrigin-RevId: 263023575