[XLA] Add a new field in HloModuleConfig to annotate shardable inputs/outputs of the entry computation.
A shardable pair of input and output that can be considered as the original and updated value of a variable maintained by the caller, and that can be sharded by XLA by creating separate sharding/unsharding programs. If sharded, the caller is responsible to call the sharding/unsharding programs before and after the main program.

PiperOrigin-RevId: 263393862