Introduce `indices_are_sorted` attribute for gather/scatter HLO
If the attribute is set to `true` then the backend can assume that the
gather/scatter indices supplied by the user are sorted what should enable
the generation of more efficient code.

If the attribute is set to `true` but the indices are not sorted then
the behavior is implementation defined.

PiperOrigin-RevId: 264574093