Add TfLite flex delegate with support for TF ops
Create a org.tensorflow.lite.flex.FlexDelegate class which
wraps its native counterpart for using TensorFlow ops in
TensorFlow Lite. Clients can either instantiate this delegate
directly when using ops that require TF ops, or add it as a
dependency to their project, and it will be instantiated when
necessary.

Also introduce a tensorflow-lite-select-tf-ops.aar target, which
is a plugin that should be used alongside tensorflow-lite.aar. The
existing tensorflow-lite-with-select-tf-ops, which is a monolithic
build that includes all of core TFLite, is now deprecated and will
soon be removed.

This work is in anticipation of pushing prebuilt
tensorflow-lite-select-tf-ops libraries to the TensorFlow Lite
Bintray repository.

PiperOrigin-RevId: 264948003