Optimize TF_Output construction when building `tf.Tensor` objects.
Each `tf.Tensor` caches a SWIG-wrapped `TF_Output` object to map that tensor to the C API's graph. We currently lazily cache this object in `tf.Tensor` on its first consumption, but it turns out that we already build the object for all tensors as a by-product of calculating tensor types. This change saves the original object in each `tf.Tensor` when it is constructed, which decreases the overhead of graph construction.

PiperOrigin-RevId: 265484399