Define ReduceWindow Op in XLA HLO dialect
ReduceWindow op takes variadic input and init_value pairs and window dimensions as the required inputs. Computation to perform in this window is modeled as a region with exactly one block. It also takes optional attributes window_strides, base_dilations, window_dilations and padding. These optional attributes have a default value depending on the input ranks if they are not specified.

PiperOrigin-RevId: 267647229