Add a XlaConfigProxy class.
Callbacks can be registered to this class so that runtime environment
flags can be parsed to change configs in the Tensorflow core. A primary use
of this is for the Tensorflow core to query the XLA JIT level, which can be
configured by some runtime environment flags in addition to ConfigProto.