Register the OpFolderDialectInterface with the tf_executor dialect
This interface is called when a constant needs to be materialized (for
instance during constant folding or canonicalization). The folder will
invoke this interface while moving up the region nesting. If the constant
comes from an expression defined within an island we don't want the constant
to be materialized at a higher level.

PiperOrigin-RevId: 268129606