Hold remote task instead of eager client in `RemoteTensorHandleData`.
The EagerContext and the cluster connections have to be mutable to survive worker failures and preemptions. When cluster changes, it's hard to update the saved `eager_client`s in all remote tensor handles. This CL changes remote tensor handle data to only hold a remote task name instead of the actual eager client. At the time when the tensor handle should be destroyed, it can easily lookup the corresponding client by task name, and prevent client program crashes if the eager clients are changed in the context.

PiperOrigin-RevId: 269678594