Quantize non-zero scalars and splat tensors
To get a fully quantized model, scalars and splats tensors are required to be
quantized. Previously, the scale to quantize scalars and splats are hard-coded
to 1.0, which causes accuracy issue for non-zero cases. This patch tries to
create a range to cover both 0.0 and the scalar value and derive the scale from
the new range. This makes sure both 0.0 and the scalar value can be repreented
by the fixed-point values.

PiperOrigin-RevId: 270784431