Replace the borrowed Node* from `NodeItem` with an int32 node_id.
This CL removes the need for an ExecutorImpl instance to retain a copy of the original Graph, which results in redundant copies of graph-related information being stored. Instead, we use the more compact NodeItem structure for performance critical operations, and fall back to calling OpKernel::def() for debug logging and producing error messages. A follow-up change will remove ExecutorImpl::graph_, which is currently still required by the DirectSession implementation.

As a bonus, this change reduces sizeof(NodeItem) from 48 bytes to 40 bytes on x86-64, by changing a 64-bit pointer to a 32-bit int and reordering the fields of NodeItem so that we don't insert padding before num_output_edges.

PiperOrigin-RevId: 271454264