Fix review comments.
Change all occurences of micro_recognition to image_recognition.
Added a line to the README.md explaining what the application does, and
that a camera module is used.
Added a .gitignore file and moved the first_10_cifar_images.h from
the .gitignore in the root directory to the newly created one.
Removed unnecessary python command from build instructions.