Relax requirement for TensorFlow BiasAdd lowering to HLO
Currently it requires input of rank at least three but that is only required for NCHW data_format. The other format NHWC supports input of rank two.

Additionally, move constraints on the lowering pattern to the BiasAdd op verification. This will reject invalid earlier in the pipeline and the verification can be shared between different backends.

PiperOrigin-RevId: 272590921