Make SavedModel deps more precise.
- The dialect itself has no dependency on any TF SavedModel files.

- tf-mlir-translate needs to depend on TF core. Otherwise, there is an error:

```
E1004 13:04:11.132277  255830 tf_mlir_translate.cc:103] Failed to load saved model
'/path/to/model.sm': Not found: No session factory registered for the given session options: {target: "" config: } Registered factories are {}.
```

The root cause is that the SavedModel loader actually needs a session available to load the model (part of the load process is to session.Run a `RestoreV2` op), along with having all the needed ops registered. That's really annoying, and I've already hit this more than once, but there doesn't seem to be a way to fix it in general, since it leads to build cycles (tensorflow uses MLIR importer). For now, all clients of the SavedModel importer will need to manually depend on "all of TF", roughly speaking.

PiperOrigin-RevId: 273602359