Add filter support for autograph
filter is one of the builtin functions in python, though it is not supported
with dataset in autograph yet (as opposed to enumerate/map).

tf.data.Dataset already have filter support so adding it makes sense I think.

This PR adds the filter support for autograph with dataset.

This PR is related to 30802 (which adds enumerate support)

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>