Make collective group resolution more robust to async misconfigurations.
In collective group resolution, each device eventually calls into
`CompleteGroupLocal`.  This function waits to get a group resolution message
from all devices in the group.  For the first n-1 devices, it checks group
consistency and enqueues the callbacks in a waiter list.  When it gets the last
group resolution call, it then invokes the callbacks on each group member.

Before this change, if any group member was configured inconsistently with the
rest of the group, such as different device type or group size, we would only
invoke the done callback on the device that caused the inconsistency.  If other
devices were enqueued in the waiter list, they would forever wait and execution
would hang.

This change stores the first error status in the `GroupRec` object.  It also
calls all waiting callbacks with this error status.  Any future group members
would also get the same error status because the `GroupRec` object, once
poisoned, would always store the error status.

This change also adds tests that cover misconfigured collective groups.

PiperOrigin-RevId: 274598770