Fuse binary ops to the following affine ops.
The binary ops include add/sub/mul/div. The RHS should be scalar constant.
The affine ops include fully_connected, conv2d and depthwise_conv2d.

When the bias is updated for the transformation, we should assign different bias dim in the filter tensor.
conv2d and fully_connected is the first dimension, and depthwise_conv2d is the last one.

PiperOrigin-RevId: 274722860