[XLA:CPU] Fix use_rdtscp condition
This was emitting rdtsc on x86 (which is inaccurate wrt out-of-order execution)
and rdtscp on all other targets (which doesn't exist). On top of that we were
using an outdated prototype of the rdtscp intrinsic.

I have the feeling that using rdtscp over rdtsc likely doesn't make a
difference, but seeing this bug didn't increase my confidence in
xla_hlo_profile results.

PiperOrigin-RevId: 275094128
Change-Id: I292d9c92fbe6d45449c7c7f095975b60d62b3e5e