Force the LSTM/GRU cell to run with implementation 1 when recurrent_dropout is enabled.
This is raised by https://github.com/tensorflow/tensorflow/issues/32987.

In short, implementation 2 was applying one dropout mask to the
recurrent state, rather than having different mask for different gates.
Updating the implementation 2 to dropout on individual gates might make
it very similar as 1, which is why we fallback 1 when recurrent_dropout
is enabled.

PiperOrigin-RevId: 275336490
Change-Id: Ib853339b1033cb4d7c343e8adb349fa9a573fdd3