Add support for replication in TPU cluster formation pass.
This adds a step where if `num_replicas` > 1 for a cluster, the cluster, wrapped in a `tf_device.launch`, is further wrapped in a `tf_device.replicate`. Associated `tf.TPUReplicatedInput` and `tf.TPUReplicatedOutput` ops of the cluster will have their respective operands and results remapped to the `tf_device.replicate`.

PiperOrigin-RevId: 276781046
Change-Id: Ifb4201382f353bfd480532baa0da65fc6932f4f7