Test mixed precision with multiple workers.
The test fails with XLA, so I disabled it when XLA is enabled. I suspect this is a bug with XLA, so I'll file a bug with them once this change is submitted.

I moved some classes from keras_test.py to test_util.py, so that they could be used in the new test as well. I also added support for fp16 losses in MixedPrecisionLossScaleOptimizer, which was already supported in the Keras LossScaleOptimizer.

PiperOrigin-RevId: 277184769
Change-Id: I702ad37d3ba5c5888a96eb2d5e796d2f8d2c08b3