[XLA] Perform the check for cross-computation buffers in alternate mem earlier.
Previously, we were performing if a definition and its use is in the same
computation only after the no-copy allocation failed. This can create illegal
references to instructions across the computation boundary. The fix is
admittedly a bit restrictive but I will investigate approaches to enable better
use of alternate memory for while loops.

PiperOrigin-RevId: 277200085
Change-Id: I4be71188a5da75d71d16353241b5622890a80c0d