Support the `inference_input_type` flag
The `inference_input_type` flag is used to set the input data types. If it is
not float, quantization is required, so this flag can override `inference_type`
when it is not float. On the other hand, if this flag is float and
`inference_type` is different from float, adaptor ops are inserted to convert
the input data to the `inference_type`.

PiperOrigin-RevId: 277337467
Change-Id: Idd355d9b794313a513439869987ce407a42f73c1