Do not hold lock for executors map while waiting for pending nodes.
It's dangerous to hold the lock while waiting for pending nodes to be executed, and the added test hangs at `ClearCaches` in context.py without this change. Deadlock happens because we hold executor_map_mu_ and waiting for pending nodes, while one of the pending nodes is a DestroyRemoteTensorHandle node, which wants to grab executor_map_mu_ when trying to access its Executor().

PiperOrigin-RevId: 277833580
Change-Id: I04eaedeb38d207dadec15998beaff991c993c2fa