Filesystem interface for modular filesystem plugins.
This is the first commit in a series of changes to add support for modular
filesystem plugins to TensorFlow. Here, we only add the interface that all
plugins must obey, as well as the checks that are being done at plugin loading
time to ensure that plugins are compatible: they implement the required methods
and the API and ABI compatibility is satisfied.

The entire implementation will reside under this experimental directory until
all filesystems that TensorFlow currently supports are ported to the modular
world. At that time, we will send a final (set of) commit(s) to migrate all
filesystems to the new world.

For details about this, consult the RFC at
https://github.com/tensorflow/community/blob/master/rfcs/20190506-filesystem-plugin-modular-tensorflow.md

PiperOrigin-RevId: 277945940
Change-Id: Iff71044622e15a07a290adf2d3c121e26451dd35