Do not cache tf.data op kernels in eager mode.
The reason for this is that tf.data op kernels may have a unique cache key every time they are executed. This will happen if the tf.data op kernel expects a user-defined function because the user-defined function is traced every time the op kernel is executed and thus the function attr will be unique. Caching such kernels provides no benefit and for programs that build input pipeline graphs in a loop results in linear memory growth.

PiperOrigin-RevId: 278697051
Change-Id: Idffc8e26c78d837671f25806b823524b904c3571