Optimize the default constructor for `tensorflow::Tensor`.
The default constructor for a tensor is primarily expensive because it has to
construct a TensorShape representing a zero-length vector. It is called in
surprisingly many places throughout the TensorFlow stack, especially when it is
necessary to receive a Tensor as an output parameter from a function. This
change adds a specialized constructor for such shapes, and plumbs it through
the Tensor constructor. By allowing the specialized constructor to be inlined, the Tensor constructor implementation becomes a few mov instructions, instead of multiple calls and branches.

PiperOrigin-RevId: 281319997
Change-Id: Ib276ec23f84feb3d5a9a400dac52b359360d73cc