Add NodeMap update for incoming control edges in ScopedAllocatorOptimizer.
Ops that are being fused by the scoped allocator optimizer may have control
edges that need to be retained once the ops are fused.  The expectation is that
in the rewired graph, the incoming control edges will be attached to the
_ScopedAllocatorConcat node, and the outcoming control edges will be attached
to the _ScopedAllocatorSplit node.

This change adds a missing update to NodeMap while rewiring control edges.  It
also adds tests for confirming correct control edge rewiring.

PiperOrigin-RevId: 281334731
Change-Id: I9b7f8ca4ccfc7d66657d0ee6a69ad5562083ccb7