Fix for minimum/maximum quantization problem
Problem description:
MINIMUM and MAXIMUM operations were not quantized properly. Specifically, the problem was that only one of the inputs was quantized while another one was left in the original data type. Because of this reason, TFLite interpreter was failing to prepare since quantization params of ops did not match.

Problem cause:
MINIMUM and MAXIMUM operators were created in the way that they only had one input, not two. Hence when looping through inputs while quantizing them, only one of two inputs was quantized.

Problem fix:
Change the definition of the aforementioned properties.

This patch contains fixes for the problem described above.

    Properties of MINIMUM and MAXIMUM operators were altered to have
    two inputs rather than one. This is safe since both 1.* and 2.* branches
    have only two inputs for these ops
    Test suite for testing Minimum and Maximum ops quantization is added
    Two small binaries have been added for testing purposes

Change-Id: Ibcbfbf7bb7118a9f2cdd8ddc03e990015ba539d7