[XLA:Python] Add Device.platform field.
This can be used to associate a Device with a particular Backend. We
use the platform string instead of having a reference to the Backend
itself since Backend in defined in Python and Device is defined in
C++.

The motivation for this is to allow JAX to derive the correct Backend
to use from a Device.

PiperOrigin-RevId: 281858391
Change-Id: If329f5723e6eefcef1bafddc53b8a58150168cf7