Apply return tuple option to entry functions only during HLO export
Currently, this option applies to all functions in the module and this differs with the logic in the existing tf2xla bridge. While creating XLA ops with computations, the existing bridge seems to be using tuple return values only for the control flow ops and the entry function. The new bridge is already using tuples for the control flow ops and in general it is better to create IR in the right format from the beginning.

This change always shortens flags and parameters naming by using 'return_tuple' instead of 'always_return_tuple' and 'use_tuple_args' instead of 'use_tuple_args_for_entry_computation'.

Separately, even the entry function return tuple handling and creating tuple in case of multiple return values should be moved outside of the translation so that the translation is simple without requiring these transformations on the IR.

PiperOrigin-RevId: 282414819
Change-Id: Ic84686bff64935d243b75e21938b5ca2c68f569c