Group variable initialization when calling lift_to_graph.
When initializing variables defined inside a @tf.function which are lifted to the outer graph, group the variables together and call lift_to_graph once.  lift_to_graph supports passing in multiple tensors and the graph to lift to is the same for all of the variable initialization.  This improves setup time.

PiperOrigin-RevId: 284242959
Change-Id: I1f84b7ac5037658b40a4a8a9a8781485b5be80df