Fix saved_model_cli tensorrt conversion
The existing saved_model_cli convert tensorrt script fails in 2.X with module
not found "tensorflow.contrib". Updated the script to use the V2 API for
TensorRT to convert a saved_model.

The max_batch_size and is_dynamic_op parameters are not valid for the V2 API
so they have been removed.