Fix failing tests that are silenced.
Some tests have been failing but the tests themselves "pass" because the wrong status is being returned. This updates the returned status to be of both FileCheck and the input pipe. Tests that are "failing" are updated. Tests that are not failing but depend on the faulty status check are also updated.

PiperOrigin-RevId: 284981762
Change-Id: Ia1f77628d49121f99327ea2859339eacd1ab36f3