Fix the test failures for requantizations
Previously, a flag `is_argument` is set to false when the operand isn't an
argument (LN 649 in the original quantization_driver.cc). This isn't correct if
the operand is the dequantized value of an argument. To make the logic clean,
the cl adds the logic to scan the argument values and intialize the states
seperately.

PiperOrigin-RevId: 286612832
Change-Id: Id44a3f09e09faa9ac6dbb7cf955b3f1d25d52150