[XLA] Use the most recent allocation in default mem instead of creating a new one.
If there is already an eviction created in the allocation sequence, and if the
new user cannot be placed in alternate memory, find the most recent default
memory allocation in the allocation sequence, and use that. This fixes a memory
corruption bug involving eviction/prefetch/prefetch sequence. This should also
get rid of redundant evictions if there has already been an earlier eviction.

PiperOrigin-RevId: 287625692
Change-Id: I6c6df1beedfee5de0537ff7aaea5ddc9cd125009