Fix various bugs in LossScaleGradientTape.
The following bugs are fixed:
1. Crash when gradients were taken w.r.t. distributed variables.
2. Crash when targets or sources were not float32.
3. Crash when some gradients were None.
4. Bogus gradients returned when loss scale was 1.

Several error messages were also greatly improved, such as when the jacobian is taken.

Changed tests to take gradients w.r.t. variables instead of tensors, as this is the more common case.

In the future, we may simply the code by having each replica have an independent loss scale. This means we do not have to synchronize the loss scales, which would greatly simplify the LossScaleGradientTape implementation, and probably improve performance as well. The downside is that not having a single loss scale may be more confusing to users. For now, I kept a single loss scale, as we have not yet decided if we want to go with this approach.

PiperOrigin-RevId: 287869750
Change-Id: Ie60aa3f9721f78ab3fe857d4fdb0d1f70fa84faf