RuntimeGraphOptimizer() used to run graph optimizer only if apply_optimizaitons or erase_inline is set. If apply_optimizaiton was off and inline function is set, then it skips graph optimization, which is wrong.
Also, this function takes separate input graph and output graph args, but
when skipping, it doesn't properly set output graph.

It's possible that the caller uses the same graph for both input and output;
hence, it maybe ok, but it's not always.

This CL checks inline_function option also. In addition, when skipping
graph optimization this CL sets the output graph with the input graph
if output graph pointer differs to the input graph.

PiperOrigin-RevId: 288553094
Change-Id: Ic115ca57437867596860369c7ada866442e4986d