core: mark function `inline` (NFC)
Mark this stub as `inline` to ensure that the body is emitted inline and
no symbol implementation is provided.  This is required as otherwise,
the function is emitted everywhere as a result of being defined in the
header, but is not given COMDAT resulting in a multiply defined symbol.

This allows linking tensorflow.dll on Windows with VS2019.