Check whether the optional actually has a value.
TriangularSolve and Cholesky use protos for configuring options. The boolean
fields in these protos are optional, and are considered to be false by default.
In our HloParser, we set the the optional<bool> storing the attribute value to
absl::nullopt. If the HLO text did not specify this attribute, we would try to
dereference this optional without checking whether it has a value.
This change adds the necessary check.
PiperOrigin-RevId: 290944536
Change-Id: Id5aaa9bc11fbbc3088eb050643e038d6aebaff0d