Split eigen_backward_spatial_convolutions_test into two.
The test file is testing two distinct headers, and in some contexts, the
complexity of this file is leading to really long compile times. By splitting in
two, we should at least enable parallel compilation if not potentially reduce
the effect of whatever behavior this is triggering in the compiler.

PiperOrigin-RevId: 291462169
Change-Id: I4df8934d8eaad1c93f986c074734eb31fa98e91a