Fix crash when float64 or mixed precision used in certain layers.
Also add many more tests to layer_corectness_test.py, so that most Keras layers are tested.

Also do not test distribution strategy without mixed precision in layer_corectness_test.py. This previuosly was only tested for ease of debugging if a test failed. But the distribution strategy tests take a very long time, so it's not worth testing with distribution strategy without mixed precision.

I made the test a py_test instead of a cuda_py_test to save a bit of GPU resources.

Fixes #35817 and fixes #35883.

PiperOrigin-RevId: 291825015
Change-Id: I8ece6de5b9d549f0de06b643774686f56775781e