Clean up OpKernel 1.
1. Clean up old code related to legacy scalars and vectors in TensorFlow. This CL does not bump the GrapDef version.
Legacy behavior is allowed for a small number of frequently used kernels and a TODO to fix this in a future CL added.
2. Move OpKernel::MakeShape into tensor_util.{h,cc}. Ideally it should go in TensorShapeUtils, but that would create a cyclical dependence.

The oldest branch of TensorFlow on github is r0.7, which is at GraphDef version 8 already, so this should have no impact on GraphDefs in the wild.

We also changed the scalar test to be strict in open source on 2017-04-03.

PiperOrigin-RevId: 292234389
Change-Id: Iee372607d9b9139d33ba7be5be8b792d9471e0f6