Addressing the RNN variable name issue.
RNN layer was historically using call() to cell since the cell is not necessary to be a layer. This assumption was quite outdated, and using call() directly will miss all the check and setting we have in __call__(). This change will ensure all the ops and variables produced for cell to have a proper name scope if the cell is a layer.

Also when build the RNN cell, adding the cell's name_scope so that it will proper show up in the variable name. This is important to avoid the naming conflict when saving eager variable in H5 format.

Fix #35731.
Fix #33776

PiperOrigin-RevId: 293023977
Change-Id: I92eeeee5aa87fefa50ff2ecc63f4d8d397ab440d