Change tfl.while to be implicit capture
Previously I had TFL While behave like a function to make it simple to export.
But this isn't really nice for transformations and resulted in needing some
patch up before export anyway. So instead, make it more optimization friendly
and instead create a pass to make it export friendly.

This partly reverts parts of flatbuffer_translate and now just exports
functions to sub graphs again.

Neither the legalization from TF while nor the pass to convert to export friendly form is added to pipeline yet. There is the possibility of having Yield op allow some shape casting to avoid having functions outlined that just invokes a function, casts and returns. But that adds additional complexity and may not be worth it.

PiperOrigin-RevId: 293500793
Change-Id: I1880116880f3f515b9e46760d630ef988e0c4954