Add pattern to legalize tf.StatefulPartitionedCall to std.call
tf.StatefulPartitionedCall has similar definition as tf.PartitionedCall,
except that tf.StatefulPartitionedCall is marked as stateful, which is
used to prevent certain optimizations (like constant folding) in TensorFlow
framework. It can also be lowered down to std.call given that the MLIR
path we use different analysis to decide whether and how to optimize.

PiperOrigin-RevId: 293899166
Change-Id: I62597eafe699a6979ab66ee082e6ff7216778bee