Removes a spurious check from AddRematerializedInstruction().
It is possible to rematerialize an instruction even if it defines a buffer that is being used by the instruction that is currently in progress. Even though that particular buffer does not help to reduce the memory usage, it may still be beneficial to rematerialize the instruction due to other buffers that it defines.

PiperOrigin-RevId: 296045046
Change-Id: I8236a04f420b341da9284b842484b815ccea4584