Allow using tf.identity to explicitly copy tensors.
When device placement policy is 'explicit', and a tensor copy is
required for an operation input, the program will fail with an error
message that suggests a solution of the form

  with tf.device(dev): x = tf.identity(x)

This does not work currently. This change adds a policy exception for
the "Identity" operation, so the above starts working.

PiperOrigin-RevId: 296317942
Change-Id: Ie2e79c8b095bed9376976a8dcd1b49a0af9b7dc9