Switch remote mirror map to key off incarnation
Since remote workers can be added and removed, using the Device pointer
is not appropriate since it is possible for Device objects to be
deleted. Thus we switch the mirrors to key of the device incarnation.

In addition we remove SetRemoteOpIdAndOutputNumToLocalTensorHandle as it
is unused, and we add some const annotations.

PiperOrigin-RevId: 296538295
Change-Id: Icdffd40a5ed242c2c0f81fbede8a35a03bd79044