Fix label_smoothing in multidimensional CategoricalCrossentropy.
When label smoothing in CategoricalCrossentropy is non-zero, it takes
tf.shape(y_true)[1] as the number of classes. However, if the true
values and predictions are multidimensional (for example when training
a POS tagger where batch elements are sentences composed of words),
a wrong value is taken and the training does not work.

This fix takes the _last_ dimension as the one containing classes.