Fix mirrored_strategy_test
There're two minor issues in the test. One is the use of virtual device, which
needs to happen at program startup, I simply removed it and use physical device instead.. Another one is an
issue with some small Tensor optimization mechanism. The cached Tensor is
always on CPU, so NCCL fails.

PiperOrigin-RevId: 296982858
Change-Id: I4753882701b7d5c35d67a847408da9d024ceb83d