Fix the BatchNorm v2 performance issue wrt "training".
Change the logic to skip the math_ops.logical_and() when layer.trainable is True, since the "training" value is only going to be changed if layer.trainable is False.

This apparently fix the regression issue (1.33x). I think the root cause might be the logical_and() with a value that is not the GPU device (layer.trainable).

PiperOrigin-RevId: 297277503
Change-Id: I9e3b6650bfa58c3d2ea8e3f023a3bd1a5160cc8d