Fix ShouldRunTest to avoid showing skipped tests as passed.
Change ShouldRunTest to return a boolean and a message to tell the reason that
the test is skipped. Change RunTest to call skipTest with the reason that the
test is skipped instead of returning a pass status.

Previously, tests that are skipped because ShouldRunTest returns false are
shown as passed in sponge logs. This change allows skipped tests shown as
skipped with a reason in sponge logs, similar to tests skipped by method
run_v2_only.

PiperOrigin-RevId: 299173826
Change-Id: I45d43c6f934ff96b7015e5920af671272fe10e26