Remove "infer" dtype policy.
It has been renamed to "_infer" and no longer is intended to be used by users. The documentation referring to it has been removed from the Policy docstring. I still have to keep it, since it is the default if a user disables V2 behavior with tf.compat.v1.disable_v2_behavior(). I documented this in the global_policy() docstring.

Since the only available Policy in TF1 was "infer", Policy is now unusable in TF1, so I now expose it in TF2 only. Even before this change, Policy was useless in TF1, since the default was already "infer".

PiperOrigin-RevId: 299918175
Change-Id: Iddd7acfb367ef69181e88d0933c7b5e657583db1