[XLA:Python] Reduce use of std::shared_ptr<> in PyLocalClient.
Instead, simplify the ownership structure: make it the responsibility of the client of the API to keep the PyLocalClient object alive.

We want to do this for two reasons:
* to enable C++ consumers of the PyLocalClient API. The excessive use of std::shared_ptr<> is an antipattern in C++ APIs.
* to allow pointers that go "up" the object tree. For example, right now it is not possible to have a pointer from a Device to its owning PyLocalClient because PyLocalClient owns Device. By having PyLocalClient be the single object that clients must keep alive, we are free to, say, have pointers between Devices and PyLocalClient in whatever arrangement we see fit.

To ensure that Python callers keep the PyLocalClient alive, we change the Python wrappers to maintain a std::shared_ptr<PyLocalClient> reference via a custom pybind11 holder type. We make PyLocalClient inherit from std::enable_shared_from_this<> to make it easier for the Python bindings to obtain a shared_ptr<> in some cases, but this does not mandate that other callers use std::shared_ptr<>s if they don't want to.

PiperOrigin-RevId: 300078385
Change-Id: Iee9dc6cab1b20a0d4bc2d3c421e90953b2af7f66