[Executor] [NFC] Avoid touching the reference count when using a const tensor as input.
This change extends the recent optimization for caching a constant output tensor from a kernel in the executor, which avoids dispatch overhead for kernels like `ConstOp`, `VarHandleOp`, etc. that always produce the same value. In this case, since the original kernel retains ownership of the tensor for at least as long as the executor is alive, there is no need to increment (and later decrement) the reference count for the underlying `TensorBuffer`, and we can avoid potentially expensive inter-core cache invalidations when those tensors are used concurrently (e.g. in inference).

PiperOrigin-RevId: 300083390
Change-Id: Ib667cb4efbdbd075ac8a43a3098dfb9f81961733