Refactor quantization selection logic into single QuantizationMode class in TFLiteConverter.
This has been a major source of confusion of all engineers reading this code and attempting to make changes in this code. This is a first stab at cleaning this up.

There is only one behavior change in this CL:
It should not be an error when supported_types=[float16, int8], it should perform int8 dynamic quantization. I removed the test cases checking for this error. This should never really be a case that the user provides to the converter anyways so its a minor point.

PiperOrigin-RevId: 300125178
Change-Id: I69ef523be14e2958a53e1a95e2c2878b068e93fd