Add Graph pruning pass to bridge pipeline.
As control rets are properly handled now, the graph pruning pass can be removed from the HLO lowering pipeline (compile_mlir_util) and moved to the bridge pass. For now, the graph pruning pass is after island coarsening due to some nodes not being reachable from a fetch/target that should not be pruned away.

PiperOrigin-RevId: 300858622
Change-Id: Ib81bc8f5285472b4a59b49f8b41c6833bc7fb14f