Update the MLIR importer and exporter to handle _output_shapes attribute.
When importing GraphDef to MLIR, remove the _output_shapes attribute. This also
fixes an issue in translating function call with the attribute to LegacyCallOp.

When exporting MLIR to GraphDef, add _output_shapes attribute to ops that
produce a ShapedType result except for the tf.Const op.

Add importer/exporter tests for _output_shapes attribute.
Modified tests that affected by this change.

PiperOrigin-RevId: 303813143
Change-Id: I1916c9fa258ce71e148137d1e5cbcc34111a37b9