Add multi worker mirrored strategy combinations to enable easier testing again
The previous was rolled back because it breaks windows builds.

The change adds has_chief and num_workers to NamedDistribution. If there're more than one workers (chief+workers), the distribute flavor of combinations library will run the test in multiple processes to simulate a multi worker setup. Users are required to call combinations.main() instead of test.main().

Note that it's the same as running the test concurrently in multiple processes. You're expected to program your test in the same as if you're writing a multi client program. There's no way to get the return value from all processes. If you need that, use multi_process_runner directly.

This is very slow at this moment. Do not use it for a large number of tests.

PiperOrigin-RevId: 305136390
Change-Id: I814323cccaee5b8b7eed3b4dfef2971ab9d09cb4