Improve error messages for invalid shapes/types.
This CL adds more context to the error message when users give us invalid tensor shapes or types.

Before, trying to use [tf.int64] as a type would produce the error "TypeError: unhashable type: 'list'". Now we will produce the error "Cannot convert value [tf.int64] to a TensorFlow DType".

Before, trying to use [[1]] as a shape would produce the error "Dimension value must be integer or None or have an __index__ method, got [1]". Now we will produce the error
"TypeError: Dimension value must be integer or None or have an __index__ method, got value '[1]' with type '<class 'list'>'

The above exception was the direct cause of the following exception:

TypeError: Failed to convert '[1]' from shape '[[1]]' to a dimension. A shape should either be single dimension (e.g. 10), or an iterable of dimensions (e.g. [1, 10, None])."
PiperOrigin-RevId: 305704020
Change-Id: I30c09faae1f44c4017ddbbb4e101c0b818dc14f1