Remove nested resource access verify
The verification was on all functions in the module. But other functions in module may still use resource variables and only those called from main func is of interest, seeing as ResourceLiftingForFunctionControlFlow would fail if it weren't able to hoist called functions from there, we can retire the verify function.

PiperOrigin-RevId: 306279514
Change-Id: I3cd62268fe99cfcffb33d2cacb5bdaad251c763b