Run "xla-legalize-tf-with-tf2xla" pass after native MLIR legalization pass
This way native legalization patterns are preferred in most cases if an op can be lowered by both the passes.

Earlier, I was thinking of having "xla-legalize-tf-with-tf2xla" first to expose bugs and whitelisting only ops that are not natively supported. There would be some maintenance overhead in this approach so seems easier to run this pass only on ops that didn't get lowered natively.

PiperOrigin-RevId: 306309276
Change-Id: Ia7ccd147ab9923afba7b7b7e9bd4eeef2fa0c021