Add xla_hlo.dynamic_reshape.
This op arises during many dynamic-shape lowerings. For example, when lowering dot_general to a single batch matmul op that has a single innermost contracting dimension, one needs to transpose/reshape all the contracting dimensions to the innermost dimension and "flatten" them to a single dimension with a dynamic reshape.

Also, move the folder and verifier of the non-dynamic reshape op to be next to each other. They somehow got separated.

PiperOrigin-RevId: 306585385
Change-Id: I0f7769a3f3086ac6c2225a9024a76fdddef26715