Rolling forward "[tf.data] Reduce locking in node processing-time calculations."
The previous version had a non-deterministic use-after-free error in "parallel_interleave_dataset_op.cc", which TSAN testing picked up. The buggy version decremented a `BlockingCounter` before calling `RecordStop()`, but the last decrement call would unblock a thread that could lead to the iterator and context being deleted before `RecordStop()` would be called. The fix is to ensure that the `BlockingCounter` is always decremented after the call to `RecordStop()`.

PiperOrigin-RevId: 306689496
Change-Id: Ic85cf7b79a96a9d3f25ca4f1043e2c82505fcca0