[TF:XLA] Avoid promoting VarHandleOp as an function argument if its initial value is not read
PromoteResourcesToArgsPass no longer adds an input argument for a VarHandleOp
resource if the initial value of the resource is not read. If the resource is
written, the pass still promotes the last live value to the output of the
function as before; this output (instead of a now omitted input argument) will
get "tf.resource_name" attribute. This change does not affect non-VarHandleOp
resource promotion.

PiperOrigin-RevId: 306694967
Change-Id: I6fa11893d1743e127092da75ba2f7bf804e3173e