Use assign_add(read_value=False) instead of assign_add().op
This is a no-op change for normal variables, but conceptually more correct. The
current variable has implemented a lazy read semantic so
assign_add(read_value=False) is the same as assign_add().op except some Python
overhead (both are AssignAddVariableOp), but this is not true for DistributedVarible yet (which is ReadVariableOp).

While we're working on making DistributedVarible behave in the same way as
normal variables, it's a fairly large change and this allows us to break into
smaller pieces.

PiperOrigin-RevId: 307102833
Change-Id: I51635ecf708a24a71f39b9247d586cb07349fa9f