[Build cleanup] Split "core_cpu_impl" into fine-grained targets (1/n).
This change starts the process of splitting up the monolithic
"//tensorflow/core/common_runtime:core_cpu_impl" target into fine-grained targets
that can be depended on individually. The change makes some small header and
forward declaration changes to avoid unnecessary transitive includes and
dependency cycles. It also revises the dependencies of tests in the
"common_runtime" directory so that they depend on fin...

PiperOrigin-RevId: 307149563
Change-Id: I68971f3c42a950e4256db1d51dd6809c9733cdbd