Extend special casing of NumPy arrays to anything supporting __array__.
If a non-tf class (np.ndarray subclass or not) has the __array__ method it is
used in combination with tf.constant to convert the input array to a tf.Tensor
before the users function or a cached graph function is called.

Note to match current tested behavior I have special cased instances of np.str_
to *not* be automatically converted to tensor.

PiperOrigin-RevId: 307375562
Change-Id: I9a84efc1dfdbb916c9894f138739e9b7e16f757e