[tf.data service] Expose number of master tasks through test server_lib.
This gives tests a way to check whether a newly started worker has registered, which is a pre-requisite for avoiding the use of sleep() in some unit tests. To remove the use of sleep(), we will also need the master to become the source of truth for whether a job has ended, instead of the DataServiceDatasetOp deciding this locally as soon as it finishes all of its current tasks.

PiperOrigin-RevId: 307902389
Change-Id: Ib0a71bb00e11b06dd3dc93306738a04f2530903e