[tf.data service] Make master the source of truth for jobs finishing.
Previously, workers would independently determine that a job has finished when all of their current tasks finished. This creates a race with adding a new task. With this change, the master becomes the source of truth for job completion. When a worker finishes a task, it notifies the master, and the master marks a job as finished when all of its tasks are finished. Once a job is finished, no more tasks can be added to it.

With this change, we can avoid a race condition in testAddWorkerMidJob, where the job might end before the worker joins.

PiperOrigin-RevId: 308116355
Change-Id: Ia575a183ce23d9a3695241f6125b8d64f1024cb0