Make shape inference on import optional
Rely on explicit shape inference pass instead and the goal is to completely remove all but the most basic shape inference from import (e.g., elemental type). As a staging step add an option to enable the shape inference on import and keep the existing the behavior until clients are all updated.

PiperOrigin-RevId: 308161697
Change-Id: I352f17e6234878d36bcda8a379492cd27da2c75e