[TF:TRT] Avoid generating segments that won't be accepted for TRTEngineOp.
When implicit batch and static engine are used, it is required that input
dimensions rather than the batch dimensions are fully specified when
contructing the TRTEngineOp. Previously, we may generate segments with inputs
that do not meet this requirement and abandon the segments when constructing
TRTEngineOp. This change avoids generating segments with dynamic non-batch
dimensions.

Add two test cases.

Modify a test that is affected by this change due to a bug in
GraphProperties::InferStatically.
PiperOrigin-RevId: 308626836
Change-Id: Ie6fc04fef43ab7a575f6a6e3761150b915dc3524