Add FunctionLibraryRuntime::RunSync() methods that avoid the need for creating a callback.
In many cases, we invoke a TensorFlow function synchronously, and have to create a callback and notification to block on the result. With some executor types (e.g. SINGLE_THREADED_EXECUTOR), this causes unnecessary atomic operations to be executed (e.g. setting and destroying the Notification), which can account for ~100ns.

PiperOrigin-RevId: 309105760
Change-Id: Ie5b3aef5c4dcd3529e6a3a2701ac152b9630cc01