[Single-threaded executor] Optimize the handling of `OpKernel::const_tensor()`.
This change applies an optimization that already takes place in the default `ExecutorState`: if a kernel produces a single constant tensor, we compute the `const Tensor*` once, and forward it to the downstream kernels without modifying the refcount. As a result, we invoke fewer virtual function calls and make fewer small temporary allocations. This makes `tf.constant()` and various DT_RESOURCE-producing kernels more efficient when using the single-threaded executor.

As an enabling step, this change re-uses the `Entry` structure from the default executor, which has support for storing a union of a (manually-constructed) `Tensor` or a `const Tensor*` as appropriate.

PiperOrigin-RevId: 309248989
Change-Id: Id0bdf19e34caee7cd5ca3ac907f4ffaa61244934