[tf.data] Make "DeleteIterator" op a `HybridAsyncOpKernel`.
The "DeleteIterator" op can block on work in the inter-op threadpool, e.g. waiting for function execution to terminate. As a result, we need to ensure that this blocking happens either on a separately created thread for the purpose (or a donated calling thread), which converting it to a `HybridAsyncOpKernel` achieves.

PiperOrigin-RevId: 309514952
Change-Id: I24f056bfbfbf4b1df0332432cc43aa96be123d6a