Fix BatchNormalization issue with virtual_batch_size when shape has None
This PR tries to address the issue raised in 32380 where
BatchNormalization with virtual_batch_size will throw out error if
shape has None:
```
TypeError: Failed to convert object of type <class 'list'> to Tensor. Contents: [8, -1, None, None, 3]. Consider casting elements to a supported type.
```

This PR converts None to -1 so that it could be passed as a tensor to `reshape`.

This PR fixes 32380.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>