Return ValueError in case of empty list input for tf.map_fn
This PR tries to address the issue raised in 39229 where
empty lists input was not checked and throw out a non-obvious error:
```python
>>> import numpy as np
>>> import tensorflow as tf
>>> fn = lambda x: x
>>> tf.map_fn(fn, [])
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Library/Python/3.7/site-packages/tensorflow/python/util/deprecation.py", line 574, in new_func
    return func(*args, **kwargs)
  File "/Library/Python/3.7/site-packages/tensorflow/python/ops/map_fn.py", line 425, in map_fn_v2
    name=name)
  File "/Library/Python/3.7/site-packages/tensorflow/python/ops/map_fn.py", line 213, in map_fn
    static_shape = elems_flat[0].shape
IndexError: list index out of range
>>>
```

In case of empty list the behavior is undefined as we even don't know the output dtype.

This PR update to perform a check and thrown out
`ValueError("elems must not be empty")` to help clarify.

This PR fixes 39229.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>