Fix enable_v2_dtype_behavior() doctest failure.
The issue is policy.py has the following in its docstring:

```
>>> # Optionaly set policy back to float32 if any other models use float32
>>> tf.keras.mixed_precision.experimental.set_policy('float32')
```

This sets global state that affects enable_v2_dtype_behavior() if run after policy.py. Specifically, the enable_v2_dtype_behavior() docstring calls disable_v2_dtype_behavior(). When disabled, the default global policy is None instead of float32. However, if the global policy is explicitly set to float32, as in the policy.py docstring, it will still be float32 even when v2 dtype behavior is disabled.

To fix, I removed the parts of the docstring that disable V2 behavior. This is an inelegant fix, so I'll try to think of a better solution later.

PiperOrigin-RevId: 310489347
Change-Id: I9186016f21874114b1d67aebe47c95e13c354cae