Fix the issue of tf.divide's return value is not a tensor
This PR fixes the issue of 39475 where tf.divide's return value
is not a tensor in case x, y in divide(x, y) are both primitive python
types.

The reason was that tf.divide relies on implict `x / y`. However,
if both x and y are not tensor, the return value will fall through
python and will not be a tensor.

This PR fixes the issue.

This PR fixes 39475.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>