Canonicalize tf.Select to tf.SelectV2.
The ops are mostly equivalent, except that Select has stricter requirements and does not support broadcasting, whereas SelectV2 does.

There is one special case to be considered in this canonicalization, which is when the predicate is a tensor and the data is multidimensional. In this case, Select op semantics dictate that the predicate tensor length must match the size of the first data dimension. This varies from normal broadcasting semantics, which are used in SelectV2, so we must reshape the tensor in this case to be compatible.

This also adds verifiers and tests for the Select and SelectV2 ops in the MLIR TF dialect.

PiperOrigin-RevId: 312362580
Change-Id: I43f326ad330c92ce279b25cecf5a2cf46714ce3f