Fix sha256sum issue of mkl_dnn/oneDNN
This PR tries to address the issue raised in 39696 where
the sha256 of the exiting workspace.bzl archive for mkl_dnn
does not match between mirrored one and the github one.

The issue comes from the fact that mkl_dnn has been renamed
to oneDNN repo, and github repackaged the archive.
The original mkl_dnn now alias to oneDNN (with re-archive).

This PR adjust the download link to go directly to
oneDNN so that the sha256 matches.

This PR fixes 39696.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>