Implement a pass that converts readonly reference variables to the corresponding resource variables.
It converts (VariableV2 -> Identity) to (VarHandle -> ReadVariable).

For the background, this pass is a part of hoisting VariableV2 ops by
re-using the pipeline for hoisting (VarHandle -> ReadVariable) cases, which
can be done by the following passes:
  - Capturing resource values into global tensors (importing saved model).
  - Promoting VarHandle ops to function input/outputs.
  - Freezing global tensor pass.

This path assumes that all the VariableV2 ops is read-only via verifying the
heuristic method that assumes that all the users of them is Identity op,
fed directly.

PiperOrigin-RevId: 312704760
Change-Id: I89ac4c0543a7954f6b27d418da63f7f1418490cd