Reduce 1-Layer Functional.__call__ overhead by ~10%.
Moves Model._set_save_spec to Layer. This allows Layer.__call__ to avoid a
hasattr check and also Model.__call__ to avoid an expensive call to a method
wrapped in trackable.no_automatic_dependency_tracking.

This should also allow SavedModel to use this spec in place of
build_input_shape in the future.

PiperOrigin-RevId: 313795786
Change-Id: Id7b23f98911468ed3e11261ac60989685de47aa1