(Mostly) Remove operator version support in TfLite Micro.
Only one op registration of a particular type can be added and that same op
registration is returned via FindOp. The version number still exists in the API
for backwards compatibility as well as compatibility with TfLite but is
irrelevant for TfLiteMicro after this change.

AddBuiltin/AddCustom API with min and max version params is removed.

This is a step towards completely removing operator versioning from TFLiteMicro.

PiperOrigin-RevId: 313882723
Change-Id: I2bfa8dcee69226d498597b6d0dbfc67c60641a48