[ROCm] Adding no_rocm tag to a unit-test regression.
The following commit introduces a new unit-test which fails on ROCm.

https://github.com/tensorflow/tensorflow/commit/dbef0933ebe4d3d85be73e88cfe5f83cac0ae1d6

The initial fix we submitted for this issue (see PR : https://github.com/tensorflow/tensorflow/pull/39914) does not seem to be a cover-all fix. It works when run locally but not so on CI servers. So temporarily putting a no_rocm tag on this test until we root cause the error on CI servers and come up with a proper fix