Always increment iterations in LossScaleOptimizer.
Now self.iterations is incremented in LossScaleOptimizer.apply_gradients even when gradients are not applied to the variables due to NaNs.

Before, self.iterations was not incremented because I considered self.iterations to represent the number of times gradients were applied to variables. But this caused confusion because apply_gradients() did not always increment self.iterations, so now it is incremented unconditionally.

PiperOrigin-RevId: 315356072
Change-Id: I4e8620d1ef84af55eb7a1f70b20671a7d7af38b9