Fix a bug in lift_to_graph.py.
For example, if a source tensor is SOME_OP:0, the SOME_OP will be added to source_ops in line 325, and also be added to op_map in line 203. However if SOME_OP has another output SOME_OP:1, it will skip the _copy_non_source(), so SOME_OP:1 won't be added to op_map. As the result, in line 349 if mutation.old_graph_tensor is SOME_OP:1, it will crash.

With this cl, all tensors including SOME_OP:1 will be added to op_map.

PiperOrigin-RevId: 315861055
Change-Id: I7aec62416051d90c37dec11a7a883452d1960c23