[tf.data] Cleanup of C++ user-defined function execution utilities.
This CL introduces a distinction between whether a function is multi-device (`is_multi_device`) and when a multi-device function backend is used to instantiate and execute the function. The difference is needed because the multi-device backend will not create a rendezvous for single device functions and if the function is not executed on CPU, it needs one.

PiperOrigin-RevId: 315998784
Change-Id: I8aebad07240298b6d21c0d7126b271f9952b264d