Parallel device: Add tests for async executors+collectives
I wasn't sure this would work rather than deadlocking. But executors are per-thread, so it looks like ParallelDevice making new threads implicitly disables async anyway (which is what we need for now).

This will deadlock if the Context's default executor is async. We do have a public API for that, so I'll fix that integration in a followup.

PiperOrigin-RevId: 316004808
Change-Id: I2c5c946557c5ff28b0fc5252850f6de5648690fd