Add fuzzer for `Status`.
This should allow us to test fuzzing infrastructure with minimal dependencies.

Since `Status` is used almost everywhere, we need to ensure that the common functionality is safe. We don't expect many crashes from this fuzzer since we only create a status and then look at the error message from it but this is a good test of the fuzzing infrastructure, with minimal dependencies (thus, it is a good test to weed out linker bloat and other linker issues).

PiperOrigin-RevId: 316163409
Change-Id: I62d192348b5f90209ce9188039e4962fe34872a8