[TF:TRT] Fix the TF-TRT bridge to respect the device assignment information in the graph nodes.
Previously, the TF-TRT bridge may put the graph nodes that are intended to run
on a non-GPU device to run with TensorRT. It may also put two graph nodes that
are intended to run on two specific but different GPUs to one TensorRT engine
which will make the two nodes run on the same GPUs. This change modifies the
bridge to only put nodes that are not intended to run on non-GPU devices to run
with TensorRT and to only group nodes with compatible device assignments into a
clust for a TensorRT engine.

Add tests to test cluster formation in the presence of device assignment. Modify
an existing test to test conversion with device assignment.

PiperOrigin-RevId: 316392620
Change-Id: I972322148db4a49ec36f9287a7780400fe815f52