Fix save model issue for stateless ConvLSTM2D layer.
The root cause is that ConvLSTM2D.state is a tuple rather than a list. When converting the state for save_model, the tuple is not converted to trackable objects since the states are (None, None). On the other hand, save_model requires all objects to be trackable when saving.

We didn't hit this issue for keras.LSTM since its state is a list, rather than tuple. The list is auto convert to ListWrapper since list itself is mutable.

This should fix https://github.com/tensorflow/tensorflow/issues/40328 and partly https://github.com/tensorflow/tensorflow/issues/38220

PiperOrigin-RevId: 317131403
Change-Id: I202d4dbdb29accc7a047d5f5a2fef08d24d05c7c