[TF2XLA] [NFC] Refactor datastructures for resource variables to not require snapshotting for compilation
Previously, `BuildXlaCompilerArguments` required taking a snapshot of all
resource variables in order to start compiling.
In this CL it can operate with a span of pointers to resource variables instead
(we CHECK at runtime that the lock is held).
That refactoring allows to launch the XLA compilation without creating an extra
reference to the underlying Tensor of the passed resource variables.

PiperOrigin-RevId: 317706126
Change-Id: I37a97601a08f165b23b4745e1b032bf91c21c313