Prepping for switch to ROCm 3.5+
Starting with ROCm 3.5 the underlying compiler used by hipcc will change from HCC to hip-clang. There will be a corresponding change in the HIP Runtime as well. This commit is part of a series which are intended to make the transition to ROCm 3.5+ easier.

ROCm 3.5+ (more specifically the hip-clang compiler) seems to be picky about not using a `__launch_bounds__` attribute value which is lower than the threads_per_warp value (64 on the ROCm platform). This change is to accomodate that pickiness.

The macro TENSORFLOW_COMPILE_IS_HIP_CLANG is only true when compiling TF with ROCm 3.5 and higher. The macro is a temporary construct to aid with the transition. Once the transition is complete, it will removed and the code updated appropriately.