Prevent data race when accessing function registry in UpdateRemoteMaster.
Currently `UpdateRemoteMaster` first get a list of registered functions, and then register them to remote workers. If functions are removed at the same time, it might hit nullptr when looking up functions by names in the previously acquired list. In this change, the function_defs come from the ToProto(). It makes a copy of the function protos to avoid the race condition.

PiperOrigin-RevId: 318166239
Change-Id: Ia13091ccc9577691c6508f8de7f7d2766ef81353