[XLA] Perform "copy protection" for aliasing on CPU and GPU
Aliasing for an HLO module is decided at compile time, but the buffer is either
donated or not at runtime.
In order to avoid recompilations for all possible aliasing configurations it is
often advantageous to opportunistically assign an aliasing which may or may
not actually hold at runtime.
In this case "copy protection" kicks in: if the aliasing specifies that the
output buffer B is aliased to input buffer A, but A is not actually donated at
runtime, we instead allocate a new fresh buffer C for the output,
and copy the contents of A into C (hence "copy protection").

PiperOrigin-RevId: 318214814
Change-Id: Ib5333aafbb5428308e15c18c950110ec6ddecdc5