Add a tool that lowers HLO to LLVM IR via XLA GPU and use that to write FileCheck tests for scatter
This will be useful to make sure we keep generating the same LLVM IR as we add an LHLO backend.

I also discovered a source of non-determinism via the scatter.hlo test, where in
IrEmitterUnnested::BuildKernelThunk we were iterating over a std::map that had
pointer keys.  Fix that as well.
PiperOrigin-RevId: 318581722
Change-Id: I0b5f0cdd760880378bb046f66468af2f0c4a6a15