Add capability to restrict functionalization via node filter
For the MLIR-based TPU bridge, functionalization runs before TPU cluster
extraction as part of the graph-to-MLIR conversion. This was problematic because
previously also non-TPU-nodes were functionalized at this stage which caused
issues in the TF v1 session runtime that assumes certain nodes are left
unchanged.
This change adds the capability to restrict functionalization to certain loops
and conditions, according to a user-defined node filter, which can be used to
fix the above issues (there will be a separate CL for this).

PiperOrigin-RevId: 318846935
Change-Id: I36078909c6091de083ffa5d57cdf63eca5f844ef