[XLA] Support short form when dumping aliasing into HLO output
Example of a short form output:

```
HloModule Module, input_output_alias={ {0}: 0, {1}: 1 }
```

This is more readable than the equivalent long form:

```
HloModule Module, input_output_alias={ {0}: (0, {}), {1}: (1, {}) }
```

and means that the output index {0} is aliased to the (entire) input 0, and the
output index {1} is aliased to the (entire) input 1.

The short format is very useful on CPU/GPU where the inputs are very rarely
tuples.

PiperOrigin-RevId: 320470986
Change-Id: I93a51d67816a1581d58ec7bf9061996133501ba1