Support duplicate functions after functionalization in signaturedef importer.
Currently control flow functionalization may generate multiple functional control flow ops that reference the same function. In signaturedef importer, the entire GraphDef is functionalized, but GraphDefImporter::Convert() is called on each sub-graph. So the conversion result for a subgraph may contain the same function in another sub-graph's conversion result.

This CL checks the duplicate functions when merging the subgraph's results to the final results.

PiperOrigin-RevId: 320504026
Change-Id: I2f9f195880f42def3a339717cded5bb6d44f2595