Fix ResourceOpLiftingPass to handle call ops that return tf.ReadVariableOp results.
It is possible for a tf.StatefulPartitionedCall to return the result of a tf.ReadVariableOp (with some potential forwarding through ops like tf.Identity). As return op operands are captured prior to replacing tf.ReadVariableOp results with function args, the new function return operands may not be correct. Instead, when replacing tf.ReadVariableOp results with function args, the operands of the new return are updated.

PiperOrigin-RevId: 320597502
Change-Id: I81f614e0b89670c978da376d5810ff82502f601f