[TF/XLA] Enable input/output aliasing in the TF2XLA bridge
The change required multiple components:

1) Switching to a different client API which takes a span of `ExecutionInput`s
(so that the buffer donation can be specified at runtime).
2) Propagating the flag to set up aliasing for resource variables.
3) Checking the reference count of updated resource variables, and transferring
the ownership at runtime if the count is one, and the aliasing specified in the
compiled module.

In order to do (3), we've needed to transfer from the model where we take a
snapshot of reference variables before the execution (and locking only for the
duration of the snapshotting) to locking the reference variables for the entire
compile+run cycle (only doable from XlaLocalLaunchBase) and instead of taking
the snapshot (requiring a copy) propagating the map to the pointers of tensors
associated with reference variables.

Testing: currently, testing was only done by inspecting the log traces
manually, and checking that the buffer was donated, the aliasing was specified
at the compile time, and the copy-protection only kicked in in cases where it
needed to.
In the future, we need to:
1) Test HLO generated from `tf.function(experimental_compile=True)`
2) Find a way to test the exact number of allocations generated during the test.
PiperOrigin-RevId: 321207626
Change-Id: I1003f90479d0f8d3cffb8aaf48b746aa1526535d