Fix bug in Return op construction while lifting resource ops.
While constructing the return op, the code assumes that the key-value pairs in SmallDenseMap `resource_arg_to_new_output` are in the same order as they are inserted. This is incorrect as SmallDenseMap container does not guarantee the ordering of key-value pairs. Instead we use the `resource_arg_to_new_output` index mapping itself to construct the return op.

PiperOrigin-RevId: 321723375
Change-Id: I85a244fcaf7d0339a1ce773b153cd4cb1af48a71