Cache first input and output in the interpreter.
Upcoming changes to memory allocations will remove the global TfLiteTensor allocation. This change prepares the interpreter by caching pointers to the first input and output tensor. This is required because all TfLiteTensor getters on the interpreter will be persistent. Calling these getters over-and-over with the same index will cause applications to run out of tail space.

PiperOrigin-RevId: 321874981
Change-Id: Ibb0add16ffd8a01cb5312d0913a96c3c69001f1f