Change Tensorboard callback to directly write out the train_function graphs at the end of the first training batch, rather than trying to write the potentially-incorrect Keras backend graph when the callback is first initialized.
Also add tests to verify that a graph containing the model actually gets written to tensorboard. It turns out it wasn't getting written for some of the sequential models in the test before this CL, because the model had yet to be called when the callback graph writing triggered, so the Keras backend graph was empty.

-----------
Note: I was looking to write the test graph too, but I started seeing errors about default callbacks not being allowed to have test batch hooks + problems w/ the validation writer being closed.
PiperOrigin-RevId: 322693354
Change-Id: I119b51841ca26c9dca14938b3e49e8f85efdb57c