Wait for eager collectives to finish before returning
The background is that we use async executors to launch collective ops in eager, so calling the strategy.reduce() in eager no longer blocks until the collective finishes. This changes that with the following motivations:

- By default the user is using a sync executor. Having strategy.reduce behave in async is inconsistent.

- This may leads to issues if strategy.reduce is the last line in the program, e.g.:

```
metrics = strategy.reduce()
if is_chief():
  write(metrics)
```

Other workers will exit before the collective finishes, and cause a deadlock or failure.

- This doesn't work with single client where collectives are all on a remote worker. Because there's no guaranteed RPC ordering between the collective op and the op that consumes the output of the collective. Note that we need to use more than one stream.

PiperOrigin-RevId: 323414702
Change-Id: I4a04d194ba53769c9df7a9ce3867eadea8dc9dea