Create a DispatcherState class for managing dispatcher state.
The goal of the class is to make it easier to keep the journal in sync with the in-memory state.

Currently only the RegisterDataset state transition is implemented. Further CLs will implement the rest of the state transitions enumerated in journal.proto, and we may add more state transitions in the future.

PiperOrigin-RevId: 323478950
Change-Id: I53433c6b2c45d01a3aabea26dcc62dae83e0af04