Lower priority of Merge nodes in LIFO scheduler
For a subgraph like Switch - Merge, without control dependency annotation,
VirtualScheduler schedules both outputs of Switch; unlike FIFO
or FirstReady, LIFO choses one output path of Switch, keep executing those nodes,
and when it reaches Merge, we schedules the Merge node (as it's Last In); then
the other path of the Switch is scheduled after the Merge; and this scheduling is
wrong in terms of control dependency. It is only a problem in VirtualScheduler,
as in real execution, only one output path of Swtich will run.

This CL puts Merge node at the beginning of the node queue in LIFOManager; that way Merge is scheduled with the lowest priority. All the inputs of Merge will be scheduled before the Merge.

It's possible that Merge is not for Switch-Merge, more commonly, Merge is for implementing while loop. When there are many NextIteration-Merge pairs, all the Merge nodes will be in the lowest priority; and it makes sense as proceeding to
the next iteration better be (loosely) sync'ed among many NextIteration-Merge pairs.

Note that it is only for LIFOManager; VirtualScheduler by default uses FirstReady, so it doesn't change the behavior unless LIFO or Composite is specifically used.

PiperOrigin-RevId: 324258814
Change-Id: Icaab672b349ce4cfd1195cf8fe10cf3b69c9fe56