[MLIR][KERNEL_GEN] Add tf framework C interface.
All alloc and dealloc ops in the kernel have to use TF OpKernelContext. std.alloc and std.dealloc are converted to tf_framework.alloc_raw and tf_framework.dealloc_raw which are lowered to external calls. This PR adds a C++ library with the external functions and also adds a fake implementation for testing.

PiperOrigin-RevId: 324998567
Change-Id: I8528ea3fdf68d0b653cc7b185563ffb50eca8571