Updating TF to acccount for the (ROCm 3.7) change in hipDeviceGetStreamPriorityRange
Starting with ROCm 3.7, the `hipDeviceGetStreamPriorityRange` API returns a range of `[-1,1]`.
This is a departure from the `[0,2]` range that was returned by this API in ROCm 3.3 and prior.

Updating the TF unit test, that has checks based on the range returned by this API, to account for change in the returned range