Parallel device: simplify saving to look like N separate variables
Also bypasses recent infra changes which mean saving needs to be trace-safe

There is some danger that we'll want eventually to run save ops on the parallel device, and that making that checkpoint-compatible with this will be hard. I think it's doable; essentially we'd be including device_id() in SaveSpec.name.

Running save ops on the parallel device needs some changes to saving infra and/or some special-casing of ops in the parallel device. We'd basically want to pass device_id() to the ShardedFilename op, and for its output to get transposed (parallel scalar -> vector) when passed to MergeV2. I doubt people replicating many ways will want to save every buffer, so probably not worth optimizing right now.

PiperOrigin-RevId: 326325670
Change-Id: I4c031ec0f3fc1e57f465e7aa3107f685e85e24f2