Bump revision of CUB we link against in OSS
This intends to unbreak the TF CUDA 11 Windows build.

CUDA 11 ships with cub and thrust.  We were pulling in a very old (~ 2018)
version of cub which was incompatible with the thrust shipped as part of CUDA
11.

To fix this, pull in the version of CUB that ships with CUDA 11.  Once TF only
supports CUDA 11+, we could remove the cub repository in workspace.bzl.

PiperOrigin-RevId: 326612096
Change-Id: I9eb3f015d473677473247b1fc89fb1014cf52f9c