Improve performance of float16 GPU ResizeBilinear.
Before there was no GPU kernel registered, so this change registers it. Unfortunately, the gradient is slow in float16 on the GPU, and also accumulates in float16 on both the CPU and GPU, so the gradient is run in float32 then casted to float16.

Fixes https://github.com/tensorflow/tensorflow/issues/41934. Thank you @edwardyehuang for discovering this issue.

Running the benchmark https://gist.github.com/reedwm/ed3fe4f2bf508a114e95b54bd39350f4 on a Titan V:

Before this change:
time for float32 (32, 16, 16, 16)->(500, 500): 0.010073661804199219
time for float16 (32, 16, 16, 16)->(500, 500): 5.622403621673584
time for float32 (32, 16, 16, 16)->(500, 500) grad: 0.05741095542907715
time for float16 (32, 16, 16, 16)->(500, 500) grad: 42.245511293411255

After this change:
time for float32 (32, 16, 16, 16)->(500, 500): 0.010007143020629883
time for float16 (32, 16, 16, 16)->(500, 500): 0.021726608276367188
time for float32 (32, 16, 16, 16)->(500, 500) grad: 0.05744624137878418
time for float16 (32, 16, 16, 16)->(500, 500) grad: 0.07420516014099121

PiperOrigin-RevId: 327680866
Change-Id: I57313bf0b0b7416b2548701c29d3ac6756dfa022