Add CheckPeerHealth to PeerAccessInterface
This method will be used to detect peer failures. It sends a GetStatus to the peer and verifies that the device incarnation matches the local record. Verifying incarnation is necessary to detect the case when a worker fails and restarts quickly.

This change also adds a GetDeviceAttributesCached method to DeviceResolverInterface in order to get the local record of device incarnation. We cannot use GetDeviceAttributesAsync because it sends a RPC with fail_fast=false if the device is not in cache. Such RPC may wait forever if the peer is down and never comes back.

PiperOrigin-RevId: 327711672
Change-Id: I870b7079b32f00803a1f509dfa009141c67fdf49