Retire MultiWorkerAllReduce
MultiWorkerAllReduce is for in graph replicated multi worker training, which is no longer supported. MultiWorkerAllReduce is no longer exposed so we shouldn't expect any usage of it.

Existing tests with MultiWorkerAllReduce are changed to use CollectiveAllReduce.
It's not officially supported but we should grandfather existing tests since
it's non-trivial to migrate them to MultiWorkerMirroredStrategy.

PiperOrigin-RevId: 328685198
Change-Id: Ib3e65fa772f05ca6a69f9e98864ac42bfdf4af51