Give custom devices the option to do type-based dispatch for ops with no explicit placement
When there is a custom device input and one or more physical device inputs to an op, presents the op to the custom device but indicates that the user did not explicitly request the placement (via the device property of the passed op).

Custom devices which want to stick to strict scope-based placement can either copy off the inputs and run the op on the default device or throw an error. The parallel device will stick with scope-only dispatch for now.

PiperOrigin-RevId: 328840123
Change-Id: Ic7490c0700a7ca5c74fd362211fa2fc9e008051c