Use check health API instead of an all-reduce
This change also introduces a barrier when creating the strategy. This implies any work after creating the strategy can only happen after the cluster is alive. This shouldn't affect training wall time, which is bounded by the slowest worker anyway.

The barrier is necessary otherwise the cluster may enter a restart storm.

PiperOrigin-RevId: 328866636
Change-Id: I74884d60fd8ff9572c6bf35ea20409ca8e66c3ae