Fix error when determining which values should be communicated to/from host.
When there is nested control flow where the highest level control flow is outside compiled, a value defined in the first level of control flow
could be captured by the nested control flow.  Since the entire control flow op including the nested op is outside compiled, the value defined in the first level of control flow does not need to be communicated to/from the host since it will live on the host.

PiperOrigin-RevId: 329962137
Change-Id: I9598af72a0693289de11d313795ae49bc1ccaf0f