Add pass that sinks resource writes (tf.AssignVariableOp) into a tf_device.parallel_execute region.
When a tf_device.parallel_execute result is only consumed by a tf.AssignVariableOp, the tf.AssignVariableOp can be move into the region, allowing for resource writes to be fused with TPUExecute in a subsequent pass. Such a case is possible with tf_device.parallel_execute being introduced via outside compilation.

PiperOrigin-RevId: 329991629
Change-Id: Iaf654b80599a4db447edf2ca6bcad868164f156b