[XLA] Handle buffer-sharing of DUS & Scatter in alias analysis.
Instead of special-casing in buffer_assignment for DynamicUpdateSlice and
Scatter to share the operand buffer with the output, we now explicitly treat
these in-place operations as a must-alias relationship in HloAliasAnalysis. This
requires inserting necessary copies in copy insertion when the operand buffer
might be used by multiple users, instead of relying on the backend to implicity
insert copies. Finally, this CL reworks the alias handling in
memory_space_assignment to more robustly propagate aliased allocation decisions
to the aliased positions.

PiperOrigin-RevId: 331352657
Change-Id: I5f95c0b1753f9a233a1b4f0ec6f7f93c87de6581