Removed micro_optional_debug_tools.
This was adding std::vector and printf to the microlite library and is one of
the reasons why micro_interpreter_test does not build for the bluepill target.

If the functionality of micro_optional_debug_tools is still needed then we can
figure out a proper way to have non-portable tools be part of the micro build.

PiperOrigin-RevId: 332125494
Change-Id: Id4b961466221861c15a7b30970b3d1e328d98857