[tf.data] Fix parallel interleave tsan failure.
The failure occurs because when restoring from a checkpoint, parallel interleave uses multiple threads to read from the IteratorStateReader, but the IteratorStateReader is not threadsafe. The IteratorStateReader is not threadsafe because it accesses its internal maps using map[key], which is non-const. This CL updates the IteratorStateReader's methods to be const, updating the map usage to only call const map functions.

PiperOrigin-RevId: 332253922
Change-Id: I5fd876c6bdbd9b668624511fbb062d6e16d441ee