In TpuCompiler::Compile, initialize TpuExecutables with correct HloModules.
Prior to this change, the compiler in tpu_on_demand_compiler.cc would
initialize TpuExecutables with the same HloModules passed to
Compile(). This isn't always correct, for example when XLA partitions
the computation and changes the input/output shapes.

This change has the compiler initialize the TpuExecutables with the
HloModules from the underlying executables returned from the TPU C
API's compile function.

Other changes included as part of this:
* Adds C API function for getting an executable's HLO module.
* Moves some functions to ApiConverter. The implementations added to
  c_api_conversions.cc are copies with no additional changes.
* Removes some duplicate function declarations from ApiConverter.
* Fixes a double-free from TpuCompiler::Compile.

PiperOrigin-RevId: 332900516
Change-Id: I7a8b6f66921259fe7c50ab1f36f4af236c77f955