Parallel device: fix variable initialization in tf.function
Switches ParallelDevice variables to be compatible with the tf.function variable creator scope, and adds a special case to handle conditional initialization of parallel variables.

Adds TPU tests for the parallel device since that's a major constraint on the implementation (no uninitialized input to tf.cond).

Rolling forward with some branching logic for Windows (may not be Windows-specific, but whatever combination of packages we test with there).

PiperOrigin-RevId: 334170699
Change-Id: I541655bd8a116d013a5a3f62b645aa7242411a40