Move devices information from instance params to group params
This removes the necessity to hold both instance_mu_ and gr->mu.

Each worker still calculates these information by themselves. This is the same as before, but it may be better if the leader passes these information to others, so that it's easy to make sure they're the same everywhere.

PiperOrigin-RevId: 334274263
Change-Id: Ic854b7459b286b7fa3fb1a4f890e0f49aed14d3b