Change SimpleMemoryAllocator to always set a specific HEAD allocation size.
The current implementation allows for EnsureHeadSize() to be called several times and ensures that the head is the size of the largest value specified. An upcoming change will reuse the head for storing kernel requested scratch buffer allocations instead of members on class (currently ContextHelper). This change allows for TFLM to adjust the head size, but always revert to the largest model head requirement when memory planning is complete for a model.

PiperOrigin-RevId: 334492115
Change-Id: Ic03e0af7b61acaccd69b2d5aeaea352d201d4c0d