If there are any variant inputs/output to outside compilation cluster, pull-in provider/consuming ops.
Variant types can not be communicated to/from TPU device.  Instead we should pull the Op providing the variant input or the op consuming the variant output of the cluster.  This is not done if the variant source/sink ops are inside of an already outside compiled region(for example control flow).

PiperOrigin-RevId: 334854448
Change-Id: Id86fc6beadd6852d68544f156c297fc816dbd84b