Add new mlir_bridge_rollout field to Config proto.
The exisiting enable_mlir_bridge field is insufficient because
it is a boolean. Either the MLIR bridge is enabled or disabled.
To slowly and safely rollout the MLIR bridge we need a third
category: unspecified. If the field is unspecified then we can
selectively choose to enable the bridge. This will allow us to
slowly enable the bridge for increasingly more models.
PiperOrigin-RevId: 335499348
Change-Id: I53dc4789c5e2aca9203e8c7dc4b0b5ca77c27f26