Add TAGS=posix to the CI build.
Additionally, PR #43571 moved the remove virtual delete macro to public from
private, presumably for the TAGS=posix build. However, that does not seem to be
reproducible and this change moves it back to private.

We are going to keep increasing the CI coverage and if CI passes then we should
be confident that things are ok.

PiperOrigin-RevId: 335685796
Change-Id: I446d9216a8e7c8b8c981d71bd0314dc7709b4e4f