[XLA:GPU] Make test case even bigger to throw off LLVM's unrolling
This test is supposed to test XLA's unrolling, but after
https://github.com/tensorflow/tensorflow/commit/b49b04b9cc921897f6707bdd7257747e70a40222
we end up with a small loop inside the kernel. The trip count of that loop
depends on the GPU architecture, making this test fail on sm_70 and up. Make
the shapes in the test really big to avoid this behavior.

PiperOrigin-RevId: 336069718
Change-Id: I073c0ab06ad20b420152958fc5f1b7e50774407d