Update NamedDistribution for lazy creation of the runner in tf.distribute.
Only create runner when the first time it is retrieved. The creation of the runner/cluster spec will trigger the portpicker code to reserve an unused port, and this is not ideal in __init__ since the creation of the strategy combination happens at tf load time, eg "import tensorflow as tf".

In the normal case, the port should only be reserved when the strategy_combination is used.

PiperOrigin-RevId: 337337130
Change-Id: I66835ada8ce7b65470c11fba92948c9caec9d970