Add option to not save the traces when exporting to the SavedModel format.
The current tracing implementation has a limited scope of models and layers that can be traced. When users add a custom layer or model that is unsupported (e.g. with multiple tensor arguments), they'll come across an error that prevents them from saving the model entirely. With this option, those models can now be saved to the SavedModel format for serving or retatining.

PiperOrigin-RevId: 337396077
Change-Id: Icfcbf35fc07cd72de53d49d29daf3db60d3dee57