Enable BufRendezvous per-entry cancellation.
This change adds functionality to cancel pending transfers in BufRendezvous via
the CancellationManager.  When a non-null CancellationManager is passed to
ProvideBuf or ConsumeBuf, the function will register a callback with the
cancellation manager.  This callback, called when cancellation starts, will
invoke the pending producer and consumer callbacks on that key in
BufRendezvous, and cleanup state as well.

This will enable finer grained cancellation of collectives.  For example, after
this change it will be possible to pass in the cancellation manager from
OpKernelContext, which will cancel pending transfers without permanently
aborting BufRendezvous.

PiperOrigin-RevId: 338122132
Change-Id: I930736ac3130878c13d29d6527960d9b74f1fb03