Refactor Keras SavedModel loading implementation.
Before this CL, Keras rebuilds from a SavedModel by overwriting steps in the internal loader, so that the loader calls a different function to load Keras layers and models. Other objects and variables are loaded using the core loading functions.

With load_partial added in cl/333129134, Keras objects and core objects can be loaded in separate steps. The Keras objects are loaded first, which are passed as inputs to load_partial, which will then load the other nodes in the SavedModel.

PiperOrigin-RevId: 338167915
Change-Id: I4d46d844552cf10d5d3cdcd32bcec60cc529c31e