Fix single pip package renaming bug
There is a bug in the single pip package renaming code: we do a replacement over the entire contents of `METADATA` and this causes the `tensorflow_estimator` dependency to be replaced with `tensorflow_gpu_estimator` (on 1.15 it was `tensorflow_cpu_estimator`). These packages don't exist by themseleves, Estimator has no CPU/GPU split. Previously this required a manual alteration of the Estimator package to fake it being the CPU/GPU version and a manual upload for that, but we should move away from this manual step as it always causes issues with new releases.

See for example #44775 (there are a few more similar issues, both internally and externally, but this is the most recent one).

We should build each pip package instead of doing the renaming. We do that on Linux/Mac already but Windows builds take too long so rather than rebuilding we just fake the new package via this renaming function. Future work in this area is needed to get rid of the renaming function, eventually removing it completely from both TF and TF ecosystem packages.

PiperOrigin-RevId: 341915841
Change-Id: I2bd4c3621e581ccf31e7bdd52958937b93971b90